:truck: Rename PDARule to PDA
